:recycle: Moved functions around
