+++ packages/reactivity/src/utils/index.ts
@@ -1,0 +1,5 @@
+export { isGetter } from './isGetter';
+export { isMapType } from './isMapType';
+export { isObject } from './isObject';
+export { isSetter } from './isSetter';
+export type { MapTypes } from './isMapType';
