+++ packages/reactivity/src/utils/isGetter.ts
@@ -1,0 +1,4 @@
+export const isGetter = (obj: object, key: string | symbol): boolean => {
+  const descriptor = Object.getOwnPropertyDescriptor(obj, key);
+  return !!descriptor?.get;
+};
