+++ packages/reactivity/src/utils/isMapType.ts
@@ -1,0 +1,5 @@
+export const isMapType = (obj: unknown): obj is MapTypes => {
+  return obj instanceof Map || obj instanceof WeakMap;
+};
+
+export type MapTypes = Map<unknown, unknown> | WeakMap<object, unknown>;
