+++ packages/reactivity/src/utils/isObject.ts
@@ -1,0 +1,2 @@
+export const isObject = (obj: unknown): obj is object =>
+  typeof obj === 'object' && obj !== null;
