+++ packages/reactivity/src/utils/isSetter.ts
@@ -1,0 +1,4 @@
+export const isSetter = (obj: object, key: string | symbol): boolean => {
+  const descriptor = Object.getOwnPropertyDescriptor(obj, key);
+  return !!descriptor?.set;
+};
